requirements.txt
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.2.0
alembic==1.11.2
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
apturl==0.5.2
argcomplete==3.0.8
asgiref==3.7.2
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bcrypt==4.0.1
billiard==4.1.0
bitarray==2.6.0
black==22.10.0
blinker==1.4
Brlapi==0.8.3
Brownie==0.5.1
cachetools==5.3.1
celery==5.3.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
command-not-found==0.3
cryptography==41.0.3
cupshelpers==1.0
cytoolz==0.12.0
dataclassy==0.11.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docx2pdf==0.1.8
duplicity==0.8.21
ecdsa==0.18.0
eip712==0.1.0
email-validator==2.0.0.post2
eth-abi==2.2.0
eth-account==0.5.9
eth-brownie==1.19.3
eth-event==1.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
exceptiongroup==1.1.2
execnet==1.9.0
fastapi==0.104.1
fastapi-jwt-auth==0.5.0
fasteners==0.14.1
filelock==3.12.0
frozenlist==1.3.1
future==0.18.2
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
hexbytes==0.2.3
httplib2==0.20.2
httptools==0.5.0
hypothesis==6.27.3
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
kombu==5.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
lockfile==0.12.2
louis==3.20.0
lru-dict==1.1.8
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
netifaces==0.11.0
numpy==1.24.0
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==23.1
pandas==1.5.2
paramiko==2.9.3
parsimonious==0.8.1
passlib==1.7.4
pathspec==0.10.2
pdf2image==1.16.3
pexpect==4.8.0
Pillow==9.0.1
pipx==1.2.0
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyasn1==0.5.0
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.15.0
pycups==2.0.1
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.13.0
pygments-lexer-solidity==0.7.0
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==1.34.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-docx==0.8.11
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.6
python-pptx==0.6.21
pythx==1.6.1
pytz==2022.7
pyxdg==0.27
PyYAML==6.0
redis==4.6.0
reportlab==3.6.8
requests==2.28.1
rlp==2.0.1
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.9.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.27.0
systemd-python==234
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
userpath==1.8.0
uvicorn==0.24.0.post1
varint==1.0.2
vine==5.0.0
virtualenv==20.23.0
vvm==0.1.0
vyper==0.3.7
wadllib==1.3.6
watchfiles==0.18.1
wcwidth==0.2.5
web3==5.31.3
websocket-client==1.2.3
websockets==10.4
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
XlsxWriter==3.1.0
yarl==1.8.1
zipp==1.0.0
